refactor(conversation):WB-3957 create useToolbarActions hook
